Add tree class, func components and main driver
Add ToTree type class and use it to print with indentation.
Start on func components. These still need locals and instructions. In addition,
components need to be sorted and put into the right entries in the components
record.
Add Main to run the Lexer or Parser on a directory of tests given flags at the
command line.